tile sprite update queue
